Refactored integration tests. extracted out common methods to single class.
