+++ application/src/it/java/com/hashmapinc/server/controller/BaseAssetControllerTest.java
@@ -18,13 +18,12 @@ package com.hashmapinc.server.controller;
 
 import com.datastax.driver.core.utils.UUIDs;
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.hashmapinc.server.common.data.Customer;
-import com.hashmapinc.server.common.data.EntitySubtype;
-import com.hashmapinc.server.common.data.Tenant;
-import com.hashmapinc.server.common.data.User;
+import com.hashmapinc.server.common.data.*;
 import com.hashmapinc.server.common.data.asset.Asset;
-import com.hashmapinc.server.common.data.id.CustomerId;
-import com.hashmapinc.server.common.data.id.DataModelObjectId;
+import com.hashmapinc.server.common.data.datamodel.AttributeDefinition;
+import com.hashmapinc.server.common.data.datamodel.DataModel;
+import com.hashmapinc.server.common.data.datamodel.DataModelObject;
+import com.hashmapinc.server.common.data.id.*;
 import com.hashmapinc.server.common.data.page.TextPageData;
 import com.hashmapinc.server.common.data.page.TextPageLink;
 import com.hashmapinc.server.common.data.security.Authority;
@@ -37,6 +36,7 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import static com.hashmapinc.server.dao.model.ModelConstants.NULL_UUID;
 import static org.hamcrest.Matchers.containsString;
@@ -52,19 +52,8 @@ public abstract class BaseAssetControllerTest extends AbstractControllerTest {
     }
 
     @Test
-    public void testSaveAsset() throws Exception {
-        Asset asset = new Asset();
-        asset.setName("My asset");
-        asset.setType("default");
-        Asset savedAsset = doPost("/api/asset", asset, Asset.class);
-
-        Assert.assertNotNull(savedAsset);
-        Assert.assertNotNull(savedAsset.getId());
-        Assert.assertTrue(savedAsset.getCreatedTime() > 0);
-        Assert.assertEquals(savedTenant.getId(), savedAsset.getTenantId());
-        Assert.assertNotNull(savedAsset.getCustomerId());
-        Assert.assertEquals(NULL_UUID, savedAsset.getCustomerId().getId());
-        Assert.assertEquals(asset.getName(), savedAsset.getName());
+    public void testSaveAndUpdateAsset() throws Exception {
+        Asset savedAsset = createAsset(null, null);
 
         savedAsset.setName("My new asset");
         doPost("/api/asset", savedAsset, Asset.class);
@@ -75,10 +64,7 @@ public abstract class BaseAssetControllerTest extends AbstractControllerTest {
 
     @Test
     public void testFindAssetById() throws Exception {
-        Asset asset = new Asset();
-        asset.setName("My asset");
-        asset.setType("default");
-        Asset savedAsset = doPost("/api/asset", asset, Asset.class);
+        Asset savedAsset = createAsset(null, null);
         Asset foundAsset = doGet("/api/asset/" + savedAsset.getId().getId().toString(), Asset.class);
         Assert.assertNotNull(foundAsset);
         Assert.assertEquals(savedAsset, foundAsset);
@@ -116,13 +102,8 @@ public abstract class BaseAssetControllerTest extends AbstractControllerTest {
 
     @Test
     public void testDeleteAsset() throws Exception {
-        Asset asset = new Asset();
-        asset.setName("My asset");
-        asset.setType("default");
-        Asset savedAsset = doPost("/api/asset", asset, Asset.class);
-
-        doDelete("/api/asset/"+savedAsset.getId().getId().toString())
-                .andExpect(status().isOk());
+        Asset savedAsset = createAsset(null, null);
+        deleteAsset(savedAsset.getId());
 
         doGet("/api/asset/"+savedAsset.getId().getId().toString())
                 .andExpect(status().isNotFound());
@@ -170,11 +151,7 @@ public abstract class BaseAssetControllerTest extends AbstractControllerTest {
 
     @Test
     public void testAssignAssetToNonExistentCustomer() throws Exception {
-        Asset asset = new Asset();
-        asset.setName("My asset");
-        asset.setType("default");
-        Asset savedAsset = doPost("/api/asset", asset, Asset.class);
-
+        Asset savedAsset = createAsset(null, null);
         doPost("/api/customer/" + UUIDs.timeBased().toString()
                 + "/asset/" + savedAsset.getId().getId().toString())
                 .andExpect(status().isNotFound());
@@ -204,10 +181,7 @@ public abstract class BaseAssetControllerTest extends AbstractControllerTest {
 
         loginTenantAdmin();
 
-        Asset asset = new Asset();
-        asset.setName("My asset");
-        asset.setType("default");
-        Asset savedAsset = doPost("/api/asset", asset, Asset.class);
+        Asset savedAsset = createAsset(null, null);
 
         doPost("/api/customer/" + savedCustomer.getId().getId().toString()
                 + "/asset/" + savedAsset.getId().getId().toString())
@@ -639,6 +613,63 @@ public abstract class BaseAssetControllerTest extends AbstractControllerTest {
 
         Assert.assertNotNull(foundAsset);
         Assert.assertEquals(2, foundAsset.size());
+    }
+
+    @Test
+    public void testPolicyForAsset() throws Exception {
+        CustomerId customerId = getCustomerIdOfCustomerUser();
+
+        loginTenantAdmin();
+        DataModel dataModel = createDataModel();
+        DataModelObject dataModelObject = createDataModelObject(dataModel);
+        Asset asset = createAsset(dataModelObject.getId(), customerId);
+        String policy = String.format("CUSTOMER_USER:ASSET?%s=%s&%s=%s:READ",
+                UserPermission.ResourceAttribute.ID, asset.getId().getId().toString(),
+                UserPermission.ResourceAttribute.DATA_MODEL_ID, dataModelObject.getId().getId().toString());
+
+        List<String> policies = Collections.singletonList(policy);
+
+        CustomerGroup savedCustomerGroup = createGroupWithPolicies(policies, customerId);
+        String getPolicyUrl = "/api/customer/group/" + savedCustomerGroup.getId().getId().toString() + "/policy";
+
+        final Map<String, Map<String, String>> displayablePolicies = doGetTyped(getPolicyUrl, new TypeReference<Map<String, Map<String, String>>>() {
+        });
+
+        Assert.assertArrayEquals(policies.toArray(), displayablePolicies.keySet().toArray());
+        Assert.assertEquals(displayablePolicies.get(policy).get(UserPermission.ResourceAttribute.ID.toString()), asset.getName());
+        Assert.assertEquals(displayablePolicies.get(policy).get(UserPermission.ResourceAttribute.DATA_MODEL_ID.toString()), dataModelObject.getName());
+        logout();
+
+
+        UserId customerUserId = getCustomerUserId();
+        assignUserToGroup(customerUserId, savedCustomerGroup);
+
+        loginCustomerUser();
+        Asset assetByCustomer1 = doGet("/api/asset/"+asset.getId().getId().toString(), Asset.class);
+        Assert.assertEquals(asset.getId().getId(), assetByCustomer1.getId().getId());
+        assetByCustomer1.setName("UpdatedAsset");
+        doPost("/api/asset", assetByCustomer1).andExpect(status().is4xxClientError());
+        logout();
+
+
+        String policyNew = String.format("CUSTOMER_USER:ASSET?%s=%s&%s=%s:UPDATE",
+                UserPermission.ResourceAttribute.ID, asset.getId().getId().toString(),
+                UserPermission.ResourceAttribute.DATA_MODEL_ID, dataModelObject.getId().getId().toString());
+        List<String> policiesNew = Collections.singletonList(policyNew);
+        updateGroupWithPolicies(policiesNew, savedCustomerGroup);
+
+
+        loginCustomerUser();
+        assetByCustomer1.setName("UpdatedAssetName");
+        Asset assetUpdated = doPost("/api/asset", assetByCustomer1, Asset.class);
+        Assert.assertEquals(assetByCustomer1.getName(), assetUpdated.getName());
+        logout();
+
+        loginTenantAdmin();
+        deleteGroup(savedCustomerGroup.getId());
+        deleteAsset(asset.getId());
+        deleteDataModelObject(dataModelObject.getId());
+        logout();
     }
 
 }
