+++ dao/src/it/java/com/hashmapinc/server/dao/service/AbstractServiceTest.java
@@ -22,11 +22,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.TextNode;
 import com.hashmapinc.server.common.data.BaseData;
+import com.hashmapinc.server.common.data.CustomerGroup;
 import com.hashmapinc.server.common.data.EntityType;
 import com.hashmapinc.server.common.data.Event;
+import com.hashmapinc.server.common.data.asset.Asset;
 import com.hashmapinc.server.common.data.cluster.NodeMetric;
 import com.hashmapinc.server.common.data.cluster.NodeStatus;
 import com.hashmapinc.server.common.data.computation.Computations;
+import com.hashmapinc.server.common.data.datamodel.AttributeDefinition;
+import com.hashmapinc.server.common.data.datamodel.DataModel;
+import com.hashmapinc.server.common.data.datamodel.DataModelObject;
 import com.hashmapinc.server.common.data.id.*;
 import com.hashmapinc.server.common.data.plugin.ComponentDescriptor;
 import com.hashmapinc.server.common.data.plugin.ComponentScope;
@@ -63,6 +68,7 @@ import com.hashmapinc.server.dao.timeseries.TimeseriesService;
 import com.hashmapinc.server.dao.user.UserService;
 import com.hashmapinc.server.dao.widget.WidgetTypeService;
 import com.hashmapinc.server.dao.widget.WidgetsBundleService;
+import org.junit.Assert;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -75,11 +81,10 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
 import java.io.IOException;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
+
+import static com.hashmapinc.server.dao.model.ModelConstants.NULL_UUID;
 
 @ActiveProfiles("dao-test")
 @RunWith(SpringRunner.class)
@@ -328,5 +333,67 @@ public abstract class AbstractServiceTest {
         }
         return new AuditLogLevelFilter(mask);
     }
+
+    protected CustomerGroup createGroupWithPolicies(List<String> policies, TenantId tenantId, CustomerId customerId) {
+        CustomerGroup customerGroup = new CustomerGroup();
+        customerGroup.setTitle("My Customer Group");
+        customerGroup.setTenantId(tenantId);
+        customerGroup.setCustomerId(customerId);
+        customerGroup.setPolicies(policies);
+        return customerGroupService.saveCustomerGroup(customerGroup);
+    }
+
+    protected void deleteGroup(CustomerGroupId customerGroupId) {
+        customerGroupService.deleteCustomerGroup(customerGroupId);
+    }
+
+
+    protected DataModel createDataModel(TenantId tenantId) {
+        DataModel dataModel = new DataModel();
+        dataModel.setName("Default Drilling Data Model1");
+        dataModel.setLastUpdatedTs(System.currentTimeMillis());
+        dataModel.setTenantId(tenantId);
+        return dataModelService.saveDataModel(dataModel);
+    }
+
+    protected DataModelObject createDataModelObject(DataModel dataModel){
+        DataModelObject dataModelObject = new DataModelObject();
+        dataModelObject.setName("Well");
+        dataModelObject.setDataModelId(dataModel.getId());
+        AttributeDefinition ad = new AttributeDefinition();
+        ad.setValueType("STRING");
+        ad.setName("attr name2");
+        List<AttributeDefinition> attributeDefinitions = new ArrayList<>();
+        attributeDefinitions.add(ad);
+        dataModelObject.setAttributeDefinitions(attributeDefinitions);
+        return dataModelObjectService.save(dataModelObject);
+    }
+
+    protected void deleteDataModelObject(DataModelObjectId dataModelObjectId) {
+        dataModelObjectService.removeById(dataModelObjectId);
+    }
+
+    protected Asset createAsset(DataModelObjectId dataModelObjectId, TenantId tenantId){
+        Asset asset = new Asset();
+        asset.setName("My asset");
+        asset.setType("default");
+        asset.setDataModelObjectId(dataModelObjectId);
+        asset.setTenantId(tenantId);
+        Asset savedAsset = assetService.saveAsset(asset);
+        Assert.assertNotNull(savedAsset);
+        Assert.assertNotNull(savedAsset.getId());
+        Assert.assertTrue(savedAsset.getCreatedTime() > 0);
+        Assert.assertEquals(asset.getTenantId(), savedAsset.getTenantId());
+        Assert.assertNotNull(savedAsset.getCustomerId());
+        Assert.assertEquals(NULL_UUID, savedAsset.getCustomerId().getId());
+        Assert.assertEquals(asset.getName(), savedAsset.getName());
+        return savedAsset;
+    }
+
+    protected void deleteAsset(AssetId assetId){
+        assetService.deleteAsset(assetId);
+    }
+
+
 
 }
